+++ src/main/kotlin/com/jaoafa/vcspeaker/stores/CacheStore.kt
@@ -110,7 +110,7 @@ object CacheStore : StoreStruct<CacheData>(
     }
 
     fun initiateAuditJob(interval: Int) {
-        timer("CacheAudit", false, 0, (1000 * 60 * 60 * 24 * interval).toLong()) {
+        timer("CacheAudit", false, 0, 1000L * 60 * 60 * 24 * interval) {
             runBlocking {
                 withData {
                     data.sortByDescending { it.lastUsed }
